requirements.txt
pytest==7.1.3
Appium-Python-Client==2.7.0